root_cause_identification/requirements.txt
python-dotenv
torch
together
faiss-cpu
pymongo
sentence-transformers
uvicorn
fastapi
pandas

